backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
requests==2.31.0
beautifulsoup4==4.12.3
readability-lxml==0.8.1
lxml==5.1.0
anthropic==0.43.0
pydub==0.25.1
python-dotenv==1.0.1
python-multipart==0.0.6
